FIX: NumPy errors in advanced examples

+++ examples/01_plotting/plot_haxby_masks.py
@@ -27,7 +27,7 @@ from nilearn.image.resampling import coord_transform
 affine = mean_img.affine
 _, _, k_slice = coord_transform(0, 0, z_slice,
                                 linalg.inv(affine))
-k_slice = np.round(k_slice)
+k_slice = int(np.round(k_slice))
 
 fig = plt.figure(figsize=(4, 5.4), facecolor='k')
